requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.0.0
pydub>=0.25.1
fpdf>=1.7.2
ffmpeg-python>=0.2.0
numpy>=1.24.0
soundfile>=0.12.1       